refactor(landing): centralize landing components from route to shared directory
- Move all components from app/(customer)/landing to components/landing
- Clean up redundant component folder in customer route
- Maintain single source of truth for landing components
- Simplify component organization
