--- src/app/(customer)/landing/Contact.jsx
@@ -1,12 +1,0 @@
-function Contact() {
-  return (
-    <section id="contact" className="min-h-screen pt-20 bg-cyan-300">
-      <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-12">
-        <h1 className="text-4xl font-bold">Contact</h1>
-        <p>List of Contact</p>
-      </div>
-    </section>
-  );
-}
-
-export default Contact;
