+++ src/app/page.jsx
@@ -1,15 +1,15 @@
 import AboutUs from "@/components/landing/AboutUs";
-import Contact from "@/components/landing/Contact";
 import Destination from "@/components/landing/Destination";
 import Intro from "@/components/landing/Intro";
 import Navbar from "@/components/landing/Navbar";
 import Package from "@/components/landing/Package";
 import Footer from "@/components/landing/Footer";
-import Filter from "@/components/landing/filter";
+import Filter from "@/components/landing/TravelRequestForm";
 import TabNavigation from "@/components/landing/TabNavigation";
-export default function Home() {
+
+export default function Dashboard() {
   return (
-    <>
+    <div>
       <Navbar />
       <Intro />
       {/* <TabNavigation /> */}
@@ -17,8 +17,7 @@ export default function Home() {
       <Filter />
       <Package />
       <AboutUs />
-      {/* <Contact /> */}
       <Footer />
-    </>
+    </div>
   );
 }
