+++ src/components/landing/Destination.jsx
@@ -1,13 +1,17 @@
 "use client";
 
 import { useState } from "react";
-import { useRouter } from "next/navigation"; // Tambahkan import router
+import { useRouter } from "next/navigation";
 import { useDestination } from "@/hooks/destination";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
+import { formatRupiah } from "@/lib/utils";
+import { FaShoppingCart } from "react-icons/fa";
+import { IoInformationCircle } from "react-icons/io5";
+import { CgCloseR } from "react-icons/cg";
 
 export default function Destination() {
-  const router = useRouter(); // Inisialisasi router
+  const router = useRouter();
   const { destinations, loading } = useDestination();
   const [searchTerm, setSearchTerm] = useState("");
   const [showAll, setShowAll] = useState(false);
@@ -22,15 +26,16 @@ export default function Destination() {
   const handleOrderNow = (destinationId) => {
     if (destinationId) {
       localStorage.setItem("selectedDestinationId", destinationId);
-      router.push("/login");
+      router.push("/booking-destination");
     }
   };
 
-  const filteredDestinations = destinations?.filter((dest) =>
-    `${dest.name} ${dest.location} ${dest.description}`
-      .toLowerCase()
-      .includes(searchTerm.toLowerCase())
-  ) || [];
+  const filteredDestinations =
+    destinations?.filter((dest) =>
+      `${dest.name} ${dest.location} ${dest.description}`
+        .toLowerCase()
+        .includes(searchTerm.toLowerCase()),
+    ) || [];
 
   const visibleDestinations = showAll
     ? filteredDestinations
@@ -49,7 +54,7 @@ export default function Destination() {
         <div className="flex justify-center mb-10">
           <Input
             type="text"
-            className="w-full max-w-md"
+            className="w-full max-w-md bg-white h-10"
             placeholder="Cari destinasi berdasarkan nama, lokasi, atau deskripsi..."
             value={searchTerm}
             onChange={(e) => setSearchTerm(e.target.value)}
@@ -62,7 +67,7 @@ export default function Destination() {
           <>
             <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-8">
               {visibleDestinations.length === 0 ? (
-                <p className="text-[#2A4D50] col-span-3 text-center">
+                <p className="text-[#2A4D50] col-span-3 text-center font-semibold text-lg">
                   Destinasi tidak ditemukan.
                 </p>
               ) : (
@@ -74,7 +79,10 @@ export default function Destination() {
                     <div className="relative h-48 overflow-hidden">
                       {dest.image ? (
                         <img
-                          src={dest.image_url || `${process.env.NEXT_PUBLIC_BACKEND_URL}/storage/images/${dest.image}`}
+                          src={
+                            dest.image_url ||
+                            `${process.env.NEXT_PUBLIC_BACKEND_URL}/storage/images/${dest.image}`
+                          }
                           alt={dest.name}
                           className="w-full h-full object-cover transform group-hover:scale-110 transition-transform duration-300"
                           onError={(e) => {
@@ -100,23 +108,29 @@ export default function Destination() {
                       <p className="text-sm text-gray-600 line-clamp-2 mb-3">
                         {dest.description}
                       </p>
-                      <div className="flex items-center justify-between">
-                        <span className="text-lg font-bold text-[#b22660]">
-                          Rp {dest.price?.toLocaleString("id-ID") || "Harga Custom"}
+                      <div className="space-y-4">
+                        {" "}
+                        <span className="text-lg font-bold text-[#b22660] block">
+                          {dest.price
+                            ? formatRupiah(dest.price)
+                            : "Harga Custom"}
                         </span>
-                        <Button
-                          className="bg-[#F0A04B] hover:bg-amber-600 text-white text-sm rounded-full"
-                          onClick={() => handleOrderNow(dest.id)}
-                        >
-                          Book Now
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          className="text-sm text-[#27445D] underline"
-                          onClick={() => handleMoreDetail(dest)}
-                        >
-                          More
-                        </Button>
+                        <div className="grid grid-cols-3 gap-2">
+                          <Button
+                            className="col-span-2 bg-[#F0A04B] hover:bg-amber-600 text-white text-sm flex items-center justify-center gap-2 cursor-pointer"
+                            onClick={() => handleOrderNow(dest.id)}
+                          >
+                            <FaShoppingCart className="text-lg" />
+                            Book Now
+                          </Button>
+                          <Button
+                            variant="outline"
+                            className="col-span-1 text-sm text-[#27445D] border-[#27445D] hover:bg-[#27445D] hover:text-white transition-colors cursor-pointer"
+                            onClick={() => handleMoreDetail(dest)}
+                          >
+                            More
+                          </Button>
+                        </div>
                       </div>
                     </div>
                   </div>
@@ -143,12 +157,14 @@ export default function Destination() {
         <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
           <div className="bg-white rounded-lg shadow-lg max-w-md w-full p-6 relative">
             <button
-              className="absolute top-2 right-2 text-gray-600 hover:text-black"
+              className="absolute top-2 right-2 text-gray-600 hover:text-black cursor-pointer"
               onClick={() => setIsModalOpen(false)}
             >
-              ✕
+              <CgCloseR size={28} />
             </button>
-            <h2 className="text-2xl font-bold mb-2">{selectedDestination.name}</h2>
+            <h2 className="text-2xl font-bold mb-2">
+              {selectedDestination.name}
+            </h2>
             {selectedDestination.image ? (
               <img
                 src={
@@ -170,15 +186,18 @@ export default function Destination() {
               />
             )}
 
-            <p className="mb-2"><strong>Location:</strong> {selectedDestination.location}</p>
             <p className="mb-2">
-              <strong>Price:</strong> {
-                selectedDestination.price
-                  ? `Rp ${selectedDestination.price.toLocaleString("id-ID")}`
-                  : "Harga dapat disesuaikan"
-              }
+              <strong>Location:</strong> {selectedDestination.location}
             </p>
-            <p className="mb-2"><strong>Description:</strong> {selectedDestination.description}</p>
+            <p className="mb-2">
+              <strong>Price:</strong>{" "}
+              {selectedDestination.price
+                ? formatRupiah(selectedDestination.price)
+                : "Harga dapat disesuaikan"}
+            </p>
+            <p className="mb-2">
+              <strong>Description:</strong> {selectedDestination.description}
+            </p>
           </div>
         </div>
       )}
