+++ src/components/landing/Package.jsx
@@ -5,12 +5,14 @@ import { useRouter } from "next/navigation";
 import { usePackages } from "@/hooks/package";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
+import { formatRupiah } from "@/lib/utils";
+import { FaShoppingCart } from "react-icons/fa";
+import { CgCloseR } from "react-icons/cg";
 
 export default function Package() {
   const router = useRouter();
   const { packages, loading } = usePackages();
   const [searchTerm, setSearchTerm] = useState("");
-  // const [visibleCount, setVisibleCount] = useState(3);
   const [showAll, setShowAll] = useState(false);
 
   const [selectedPackage, setSelectedPackage] = useState(null);
@@ -24,25 +26,20 @@ export default function Package() {
   const handleOrderNow = (packageId) => {
     if (packageId) {
       localStorage.setItem("selectedPackageId", packageId);
-      router.push("/login");
+      router.push("/booking-package");
     }
   };
 
-  const filteredPackages = packages?.filter((pkg) =>
-    `${pkg.name} ${pkg.description}`.toLowerCase().includes(searchTerm.toLowerCase())
-  ) || []; // ← tambahkan fallback array kosong jika undefined
+  const filteredPackages =
+    packages?.filter((pkg) =>
+      `${pkg.name} ${pkg.description}`
+        .toLowerCase()
+        .includes(searchTerm.toLowerCase()),
+    ) || []; // ← tambahkan fallback array kosong jika undefined
 
   const visiblePackages = showAll
     ? filteredPackages
     : filteredPackages.slice(0, 3);
-
-
-  // // Potong hasil yang ditampilkan berdasarkan visibleCount
-  // const displayedPackages = filteredPackages?.slice(0, visibleCount);
-
-  // const handleLoadMore = () => {
-  //   setVisibleCount(filteredPackages.length);
-  // };
 
   return (
     <section
@@ -61,7 +58,7 @@ export default function Package() {
         <div className="flex justify-center mb-10">
           <Input
             type="text"
-            className="w-full max-w-md"
+            className="w-full max-w-md bg-white h-10"
             placeholder="Cari berdasarkan nama atau deskripsi paket..."
             value={searchTerm}
             onChange={(e) => setSearchTerm(e.target.value)}
@@ -74,7 +71,7 @@ export default function Package() {
           <>
             <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-3 gap-8">
               {visiblePackages?.length === 0 ? (
-                <p className="text-white col-span-3 text-center">
+                <p className="text-white col-span-3 text-center font-semibold text-lg">
                   Paket tidak ditemukan.
                 </p>
               ) : (
@@ -86,7 +83,10 @@ export default function Package() {
                     <div className="relative h-48 overflow-hidden">
                       {pkg.image ? (
                         <img
-                          src={pkg.image_url || `${process.env.NEXT_PUBLIC_BACKEND_URL}/storage/images/${pkg.image}`}
+                          src={
+                            pkg.image_url ||
+                            `${process.env.NEXT_PUBLIC_BACKEND_URL}/storage/images/${pkg.image}`
+                          }
                           alt={pkg.name}
                           className="w-full h-full object-cover transform group-hover:scale-110 transition-transform duration-300"
                           onError={(e) => {
@@ -111,24 +111,26 @@ export default function Package() {
                       <p className="text-sm text-gray-600 line-clamp-2 mb-3">
                         {pkg.description}
                       </p>
-                      <div className="flex items-center justify-between">
-                        <span className="text-lg font-bold text-[#b22660]">
-                          Rp {pkg.price.toLocaleString("id-ID")}
+                      <div className="space-y-4">
+                        <span className="text-lg font-bold text-[#b22660] block">
+                          {formatRupiah(pkg.price)}
                         </span>
-                        <Button
-                          className="bg-[#F0A04B] hover:bg-amber-600 text-white text-sm rounded-full"
-                          onClick={() => handleOrderNow(pkg.id)}
-                        >
-                          Book Now
-                        </Button>
-                        <Button
-                          variant="ghost"
-                          className="text-sm text-[#27445D] underline"
-                          onClick={() => handleMoreDetail(pkg)}
-                        >
-                          More
-                        </Button>
-
+                        <div className="grid grid-cols-3 gap-2">
+                          <Button
+                            className="col-span-2 bg-[#F0A04B] hover:bg-amber-600 text-white text-sm flex items-center justify-center gap-2 cursor-pointer"
+                            onClick={() => handleOrderNow(pkg.id)}
+                          >
+                            <FaShoppingCart className="text-lg" />
+                            Book Now
+                          </Button>
+                          <Button
+                            variant="outline"
+                            className="col-span-1 text-sm text-[#27445D] border-[#27445D] hover:bg-[#27445D] hover:text-white transition-colors cursor-pointer"
+                            onClick={() => handleMoreDetail(pkg)}
+                          >
+                            More
+                          </Button>
+                        </div>
                       </div>
                     </div>
                   </div>
@@ -146,20 +148,21 @@ export default function Package() {
                 </Button>
               </div>
             )}
-
           </>
         )}
       </div>
+
       {isModalOpen && selectedPackage && (
         <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50">
           <div className="bg-white rounded-lg shadow-lg max-w-md w-full p-6 relative">
             <button
-              className="absolute top-2 right-2 text-gray-600 hover:text-black"
+              className="absolute top-2 right-2 text-gray-600 hover:text-black cursor-pointer"
               onClick={() => setIsModalOpen(false)}
             >
-              ✕
+              <CgCloseR size={28} />
             </button>
             <h2 className="text-2xl font-bold mb-2">{selectedPackage.name}</h2>
+
             {selectedPackage.destination?.image ? (
               <img
                 src={
@@ -180,15 +183,26 @@ export default function Package() {
                 className="w-full h-48 object-cover rounded mb-4"
               />
             )}
-            <p className="mb-2"><strong>Destination:</strong> {selectedPackage.destination?.name || "N/A"}</p>
-            <p className="mb-2"><strong>Location:</strong> {selectedPackage.destination?.location || "N/A"}</p>
-            <p className="mb-2"><strong>Duration:</strong> {selectedPackage.duration}</p>
-            <p className="mb-2"><strong>Price:</strong> Rp {selectedPackage.price.toLocaleString("id-ID")}</p>
-            <p className="mb-2"><strong>Description:</strong> {selectedPackage.description}</p>
+            <p className="mb-2">
+              <strong>Destination:</strong>{" "}
+              {selectedPackage.destination?.name || "N/A"}
+            </p>
+            <p className="mb-2">
+              <strong>Location:</strong>{" "}
+              {selectedPackage.destination?.location || "N/A"}
+            </p>
+            <p className="mb-2">
+              <strong>Duration:</strong> {selectedPackage.duration}
+            </p>
+            <p className="mb-2">
+              <strong>Price:</strong> {formatRupiah(selectedPackage.price)}
+            </p>
+            <p className="mb-2">
+              <strong>Description:</strong> {selectedPackage.description}
+            </p>
           </div>
         </div>
       )}
-
     </section>
   );
 }
